Using temp directory for profiles
